+++ CHANGELOG.md
@@ -29,17 +29,19 @@
   - (Fix) Accessing `BoundingRectangleMillimeters` was not calculating the `BoundingRectangle` if necessary
   - (Fix) SL1: Files with thumbnails size different from original SL1 would not display under the UI
   - (Fix) Anycubic: Model information is not set after an auto-conversion from SL1
+  - (Fix) CTB: Error when open file in partial model with malformed file due invalid checksum
   - (Fix) Image files: Could not find a writer for the specified extension (#755)
 - **UVtoolsCmd:**
   - (Add) "compare" command to compare two files and output the differences
   - (Add) "print-formats" command to print the available formats
+  - (Improvement) Improve "extract" command to allow extract specific thumbnails or layers in a range or indexes (#754)
 - **Project:**
   - (Change) Rename UVtools.WPF to UVtools.UI
   - (Change) Make UVtools.UI nullable enabled
   - (Remove) UVtools.GUI
   - (Improvement) Use CompiledBinding instead of ReflectionBinding
   - (Improvement) Sign the libraries and project with a strong key
-  - (Upgrade) .NET from 6.0.18 to 6.0.21
+  - (Upgrade) .NET from 6.0.18 to 6.0.22
   - (Upgrade) AvaloniaUI from 0.10.21 to 11.0.4
   - (Upgrade) OpenCV from 4.7.0 to 4.8.0
 - (Add) PrusaSlicer printer: Elegoo Saturn S